refactor: complete test suite structural alignment
- Moved GeminiConstantsSpec to model/domain
- Extracted ApiKeySpec to config package
- Removed empty algebra test directory
- Verified all tests pass
