+++ src/test/scala/gemini4s/model/domain/GeminiConstantsSpec.scala
@@ -1,9 +1,10 @@
-package gemini4s
+package gemini4s.model.domain
 
 import cats.effect.IO
 import fs2.Stream
 import munit.CatsEffectSuite
 
+import gemini4s.GeminiService
 import gemini4s.error.GeminiError
 import gemini4s.model.domain._
 import gemini4s.model.request._
